refactor: streamline comments module and remove unused mappers
- Removed the comments.mapper file and consolidated mapping functions into a new index file for better organization.
- Updated commentsController to utilize the new toCommentsResponse function for improved response handling.
- Adjusted imports in comments.service.ts to reflect the new structure, enhancing maintainability.
