+++ src/articles/mappers/to-create-article-input.mapper.ts
@@ -1,0 +1,13 @@
+import type { CreateArticleDto } from '../dto';
+import type { CreateArticleInput } from '../interfaces';
+
+export function toCreateArticleInput({
+  article,
+}: CreateArticleDto): CreateArticleInput {
+  return {
+    title: article.title,
+    description: article.description,
+    body: article.body,
+    tagList: article.tagList ?? [],
+  };
+}
