+++ src/articles/mappers/to-feed-domain.mapper.ts
@@ -1,0 +1,15 @@
+import type { ArticleFeedRow, IArticleFeed } from '../interfaces';
+
+export function toFeedDomain(article: ArticleFeedRow): IArticleFeed {
+  return {
+    slug: article.slug,
+    title: article.title,
+    description: article.description,
+    tagList: article.tagList,
+    createdAt: article.createdAt,
+    updatedAt: article.updatedAt,
+    favorited: article.favorited,
+    favoritesCount: article.favoritesCount,
+    author: article.author,
+  };
+}
