+++ src/articles/mappers/to-new-article-row.mapper.ts
@@ -1,0 +1,13 @@
+import { slugify } from '@/utils/slugify';
+import type { CreateArticleInput, NewArticleRow } from '../interfaces';
+
+export function toNewArticleRow(
+  input: CreateArticleInput,
+  authorId: number,
+): NewArticleRow {
+  return {
+    ...input,
+    slug: slugify(input.title),
+    authorId,
+  };
+}
