+++ src/comments/dto/index.ts
@@ -1,2 +1,3 @@
 export * from './create-comment.dto';
 export * from './comment-response.dto';
+export * from './comments-response.dto';
