+++ src/comments/mappers/index.ts
@@ -1,1 +1,3 @@
-export * from './comments.mapper';
+export * from './to-comment-response.mapper';
+export * from './to-comments-response.mapper';
+export * from './to-new-comment-row.mapper';
