+++ src/comments/mappers/to-comment-response.mapper.ts
@@ -1,5 +1,4 @@
 import type { CommentResponseDto } from '../dto/comment-response.dto';
-import type { NewCommentRow } from '../interfaces';
 import type { IComment } from '../interfaces/comment.interface';
 
 export const toCommentResponse = (comment: IComment): CommentResponseDto => ({
@@ -16,15 +15,3 @@ export const toCommentResponse = (comment: IComment): CommentResponseDto => ({
     },
   },
 });
-
-export function toNewCommentRow(
-  body: { body: string },
-  articleId: number,
-  authorId: number,
-): NewCommentRow {
-  return {
-    body: body.body,
-    articleId,
-    authorId,
-  };
-}
