+++ src/comments/mappers/to-new-comment-row.mapper.ts
@@ -1,0 +1,13 @@
+import type { NewCommentRow } from '../interfaces';
+
+export function toNewCommentRow(
+  body: { body: string },
+  articleId: number,
+  authorId: number,
+): NewCommentRow {
+  return {
+    body: body.body,
+    articleId,
+    authorId,
+  };
+}
